+++ ChangeLog.md
@@ -1,9 +1,12 @@
 # Change log
 
+## `v0.2.2`
+
+- Add Cached stores
+
 ## `v0.2.1`
 
 - Improve error messages
-- Add Cached stores
 
 ## `v0.2.0`